pyproject.toml
[tool.poetry]
name = "pymatrixbot"
version = "0.1.0"
description = ""
authors = ["jonmatthis <jonmatthis@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
simplematrixbotlib = "^2.11.0"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
